+++ client/src/app/shared/models/notification-data.model.ts
@@ -1,0 +1,6 @@
+export interface NotificationData {
+    title: string;
+    text: string;
+    linkTitle: string;
+    link: string;
+}